// src/json/schema.ts
export enum DocNodeKind {
  file = 0,
  function = 1,
  functionSignature = 2,
  enum = 4,
  enumMember = 8,
  class = 16,
  typeInterface = 32,
  typeAlias = 64,
  variable = 128,
  type = 256,
  simpleType = 257, // type | 1
  tupleType = 258, // type | 2
  objectType = 260, // type | 4
  functionTypeDocNode = 264, // type | 8
  property = 512,
  index = 1024
}

export interface DocNode {
  /**
   * May be the empty string or __unknown if not known.
   */
  name: string
  kind: DocNodeKind
}
// ...
export interface DocNodeComment {
  comment: string
  tags: DocNodeTag[]
}

/**
 * jsdoc tags, with the exceptions of @property and @param
 */
export interface DocNodeTag {
  tagName: string
  comment: string
}
// ...
export interface EnumDocNode extends DocNode {
  kind: DocNodeKind.enum
  jsdoc: DocNodeComment
  const: boolean
  members: EnumMemberDocNode[]
}

export interface EnumMemberDocNode extends DocNode {
  kind: DocNodeKind.enumMember
  jsdoc: DocNodeComment
  type: 'string' | 'number'
  value: string
}